requirements.txt
alembic==1.2.1
astroid==2.3.1
autopep8==1.4.4
awscli==1.16.213
botocore==1.12.203
certifi==2019.6.16
chardet==3.0.4
Click==7.0
# dlib==19.17.99
docutils==0.14
entrypoints==0.3
face-recognition==1.2.3
face-recognition-models==0.3.0
flake8==3.7.8
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.4
Pillow==6.1.0
pyasn1==0.4.6
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.9
starlette==0.12.8
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.2
wrapt==1.11.2
gunicorn==19.9.0
celery==4.3.0
Flask==1.1.1
Flask-RESTful==0.3.7
psycopg2-binary==2.8.3
urllib3==1.25.3
uvicorn==0.9.0
uvloop==0.13.0
websockets==8.0.2
Werkzeug==0.15.6
